refactor: correct preedit view behavior and polish its appearance
- Fix the view wouldn't disappear after changing theme or color on site
- Allow to custom its background color
- Add top left corner radius
- Improve its visual effect of showup
- Get alpha value as float

+++ app/src/main/java/com/osfans/trime/ime/composition/ComposingPopupWindow.kt
@@ -55,7 +55,8 @@ class ComposingPopupWindow(
                     theme.generalStyle.layout.border,
                     "border_color",
                     theme.generalStyle.layout.roundCorner,
-                    theme.generalStyle.layout.alpha,
+                    theme.generalStyle.layout.alpha
+                        .toInt(),
                 ),
             )
             width = WindowManager.LayoutParams.WRAP_CONTENT
